Introduced positional parameters for QL
This makes QL *vastly* more powerful, with a very modest enhancement to
the language.  Technically, this is a Binding, but a special one.

Renamed "QL Expression Type" to the clearer "Function".  It's an
important enough concept that I'm willing to carve out the namespace for
it, and it now works as such.