Bugfix: _sort was failing on Text types
Surprising that I hadn't caught this before, but _sort would crash
internally if you tried to sort on Text or Large Text.  The reason is
that, by the time the property's value gets to _sort, it has usually
been rendered to ParsedTextType -- which didn't have an explicit
comparer *and* is deliberately not serializable, so the default wouldn't
work.

Fix was trivial -- just add a proper comparer to ParsedTextType.  Added
a couple of unit tests to check this out.