Enhance _sort to cope with empty values
Realized that the requirement that _sort fails on empty is a pain in the
ass in practice.  So added the concept that, if the sort key is empty,
we just use the expected Type's default instead.  This will generally
cause the empties to clump together, which is what we want.