Display the page's OID in small print
When looking at a Thing, we now display its OID, in ThingID format, in
the subtitle line.  This is clickable, and provides a way to get to the
perma-link to this Thing.

This is a precursor task to making Name derived from Display Name, which
will (I believe) tend to make Names less stable.