Prevent the creation of a blank Display Name
Alexx accidentally created a Model with a blank (not None -- an empty
String) Display Name, which turns out to be easy to do and terribly bad.
Once a Thing has a blank Display Name, it becomes very hard to get to,
because all links to it wind up zero-length.

Fixed by changing Display Name to a new Type, NonEmptyPlainTextType.
This is simply PlainText with an additional constraint, that the length
must be greater than zero.