Better but more complex solution to blank Display Name
Decided that the previous fix was hackish and heavy-handed, so undid it
and instead implemented a much more elegant solution.  There is now a
Minimum Text Length meta-Property, which gives the minimum length that
values of this Property can be.  In Display Name, that is set to 1.

This touched a lot of code, mostly because I did it right.  Validate now
receives a lot more contextual information, so it will be much easier to
do this sort of validation meta-property in the future.