Remove the annoying Editor popups
There have been popups in the Editor for ages, which say whether the
value you are hovering over is inherited or overridden.  I have
concluded that that popup is hellishly annoying, and it is causing
actual problems on some mobile browsers.  (See bug .3y284t6)  So I'm
just removing it.