Bug .3y284r2 redux (all-spaces Display Name)
Alexx pointed out that a Display Name that consists only of spaces still
has all the unreachability problems.  So redefined Minimum Text Length
to be the *trimmed* length of the Text, which typically is what you care
about.

This is actually a much smaller change than it looks (one word), but
moved the new Property out to a Module, to encourage the process of
brushing stuff out of core.  The new Types Module will eventually absorb
all Type-oriented code, including Type Validation Properties.