Bug .3y284o9: Make inherited values display properly in Editor
Turns out to be a persnickety change, but mainly introduced the
"effectiveV" in DisplayPropVal -- either the local value *or* the value
that is inherited by the parent.  And then stop intentionally zero'ing
out the values of the inherited displays.

Also, made Display Name not inherited, since this change otherwise tends
to make it show up.  This also has the advantage of immediately putting
Display Name into required mode if it is specified.