Bug .3y284zh et al: fixed edge cases without Display Names
The bug was reported that Properties would lose their Names entirely if
you edited them.  This was true, but the actual problem was broader: in
any case, if you removed the Display Name, it would lose the Name.

Fixed that, and made it official that Properties use Name, *not* Display
Name.  For now, I think that's the right rule.

Removed the over-wrought meta-Property NotDisplayedInInstancesProp, and
instead simple special-cased NameProp, since I think it *is* a pretty
weird exception.