More fixes for New Tag Set Type
When you reify a Tag, it now copes with ones defined in New Tag Sets.

Hacked a variant of AsName, so that TagThing can produce URLs that don't
mangle dashes and spaces.

Editor now lets you set a Link Model on a New Tag Set.

Fixed a bug in _tagRefs.