Introduced initRequires()
This is a cleaner version of dependsUpon than the usual Ecology way.
Usually, we need an import, an explicit dependsUpon reference, and a get
-- this redundancy is a source of errors.  But by pushing Scala to the
hilt, we can be more DRY about it, with a single reference that
implicitly declares the dependency, and provides access to the interface
once it is ready.