Laced the Ecology into Thing itself
I do this a bit reluctantly, since it costs four bytes per Thing, but I
think it's the only way to deal cleanly with a lot of problems, and it
lets us stop hacking and working around.

The price is more places that explicitly refer to ecology.Ecology, but
most of those are Things that need to get transferred into Ecots.  So
this is incentive to move along on that.

The biggest general approach change here is that the Ecology is now
*implicit* in many places.  So we don't have to call it out very often:
it just permeates the atmosphere.