Lifted Collections into Core
Checkpoint.  This caused all hell to break loose, unsurprisingly, and we
wound up touching most files again.  At the moment, we compile, but a
bunch of tests fail because of a dependency loop in YesNoType.

Commented out export for now -- it's broken, and never worked yet
anyway.

NameCollection is not yet lifted, but still needs to be.

A bunch of stuff from Thing was lifted into SpacePersistence, which is
where it belonged in the first place.  Ditto for the "unresolved"
mechanism.

Renamed the temporary Ecology static to "theEcology", to clarify some
name collisions with the type.

Pulled CoreEcot out of QuerkiEcot, since Core is weird and crucial.
This allows all QuerkiEcots to not just depend on Core but see it by
default, which makes sense.  (Since Core has the tools that everything
else needs.)

Moved some QL-generated Value types out of QValue into QLEcot, which
seems appropriate.

Made QLParser Ecology-aware.

Pulled the base classes for methods into a new MethodDefs mix-in, which
Ecots can use if they need it.  This seems a good compromise in terms of
verbiage and dependency.

Lifted Evolutions into an Ecot.