Moved remaining Types to Core
TagSetType also moved, to Tags.  Also pulled UnknownNameType out of
QValue, into QL.

Started out moving just NameType, but that was the final straw -- had to
move IntType and YesNoType as well, because NameType is the absolute
center of the dependency onion.

Had to move NameCollection as well, and finally kill off the old setName
mechanism in favor of something more reliable.  Note that setName is now
accessed through Core.

Moved the static functions of NameType into NameUtils.

Wound up pulling out the nice extras in YesNoType's companion into
logic.YesNoUtils, which mixes in to Ecots.