Experiment to prove that lazy vals are okay in case classes
I had realized that I was making an important assumption: that when you
copy a case class, it would re-evaluate the lazy vals.  That seemed
reasonable, but was too important an assumption to leave untested.

Fortunately, it works.  This will probably have important ramifications
for optimizing the usage of SpaceStates.