Initial unit-testing of save/load
Not doing an *awful* lot yet, but we now have the framework for testing
serialization and deserialization, which is going to be very important
this week.

Finally introduced the notion of stubbing Ecots that need it.  Stubbed
out UserAccess, since that is totally DB-centric, and introduced the
beginnings of a testable stub for it.