Moved _instances to the new style
This proved to be the final straw for a lot of code -- getting this
working wound up moving a great deal of mid-level code to the new
Invocation style.

Invocation now includes the definingContext, if any, so we can begin to
move away from the old style for that.

Did enhance _instances to deal with received Lists, so it is now more
correct.