Pass rc down the input-render pathways, instead of state
Checkpoint.  We're going to need the rc in order to render an edit
control.  It's kind of surprising that we got this far without doing
this.  Note that this is a strict superset of the data we were
previously passing around: rc contains the state.  (And in fact, state
always came from it in the first place.)