Editing of Model List values mostly working
Checkpoint -- lots more testing to do.  But it is now shading into
"mostly working".

Horrible hack to deal with the fact that createThing returns a ThingId,
but expects to receive an OID parameter; this was causing crashes when
we were creating a Type and then a Property.

Renamed "Properties to Edit in Instances" to simply "Instance
Properties".  Made a bunch of ModelType-related functions respect that:
it governs which props get displayed and serialized, as well as edited.
A Model is now only allowed to become a Type if it has Instance
Properties defined.

Fixed several bugs in both client and server, relating to Model Lists.

Introduced the notion that, in a List's input template (which is copied
for when you create a new element), we set the index to -1 to signal
that it's a copy.  This lets us do string replacement in the various
properties when we copy it.

QLog now exposes renderBundle(), for when you want to display a
PropertyBundle in a non-spew situation.