High-level changes to Model Lists now seems to work
So far, so good.  The change needed to make the major changes (adding,
deleting and rearranging the List of Model Values) was first, on the
client, to wipe the property entirely before starting, serializing all
of the fields in the remainder, and then, on the server, building up the
new Thing change-by-change.

Needs more testing, but seems to work.