Fixed a bug in HTML rendering
HtmlRenderer was using xml.toString where it should have been using
Xhtml.toXhtml().  The result was that, when there was an embedded empty
textarea (as in the inputTemplate of a Model Type), it was rendering
illegally in HTML, and causing browser-side problems.

Also, removed the Summary from the base Property, so that text doesn't
show up for all the other Properties.