Bugfix in editing Lists of Model Values
When you created a new element of a Model Value, if that contained a
textarea (that is, a Large Text), it wasn't getting fixed up.  This was
because the code was only fixing up "input"s, where it should have been
fixing up ".propEditor"s.

Also, a bit of refactoring of the wikitext-to-xml link, since that is
becoming more common.