Fixed several bugs
First, Not Inherited was not showing up in the UI.  For some reason, it
was being treated as Internal.  I am not sure why, and slightly
disturbed by this.

Introduced Thing.firstOpt(), which I should have done ages ago.

Fixed a couple of bugs in display of Links without Link Model.

Enhanced QLog.renderBundle() to make it more useful.