.3y2855h: Enhanced Manifest to invalidate cache if new value added
Added a concept of cache invalidation to MarcoPolo, and added a flag to
Manifest saying whether we have accepted an "arbitrary" value.  If so,
we invalidate the cache, because the new value is probably a new Tag
that we should be picking up from here on out.

Also, fixed a horrible (and possibly long-standing?) bug, that you can't
delete the *last* Tag from a Tag Set in the live Editor.