.3y285b6 and .3y28573: local Properties on Instances work better now
Started out fixing the first one -- the fact that you couldn't see
Instance-local Properties if Instance Properties was set in the Model.
This revealed the latter: that you couldn't *delete* Instance-local
Properties in that situation.  Both of these are now fixed: the
Instance-local Properties are always shown in the Editor, and they can
be deleted.

Along the way, introduced the querki.test.logEdits config flag, for
debugging.