.3y285bd et al: bugs in Lists and Tags
Fixed this bug -- a Tag Set contained in a List of Model Values was
showing up in immense font.  This was just a dumb CSS bug.

Also -- if you had a Tag Set in a List of Models, when you instantiated
a new Model Value, the Tag Set came up uneditable until you refreshed.
This was a subtle initialization problem, and required some tweaking, to
not initialize Manifest until we actually instantiated the field.