Replaced SpaceManager object with an Ecot
The way we expose the SpaceManager ActorRef is a long-standing gripe
with the code.  It is now hidden behind an Ecot, where it belongs, which
means that consumers of it only refer to the SpaceOps interface instead
of the implementation of SpaceManager.

This is a step in the right direction, but we should generalize the way
that we actually create these Manager Actors from the Ecology.