Actors are now owned by Ecots
Specifically, the SpaceManager is now created by the SpaceEcot.  It
lives under QuerkiRoot in the hierarchy, but the creation is owned in
the right place.  This should make a better model for the new UserCache
I want to introduce.

Along the way, changed the SpacePersistenceFactory to an Ecot, which it
really should always have been.