.3y285gn: shouldn't show Conversations on an unreified Tag
The bug is actually a side-effect -- some complex interaction of trying
to show conversations illegally, which is triggering some sort of
attempt to show the Login screen (presumably because you're not allowed
to look at UnknownOID), which is hitting security boundaries between
http and https on the production server.  But the underlying cause is
trying to show Conversations where they make no sense anyway.