Fixed so that Large Text fields in Lists of Model Types autosize right
This one was subtle (found by Eric): when you added a new Model Value to
a List, if it contained a Large Text, that Large Text didn't autosize
until you refreshed the page.

The issue was that autosize() can only be applied to an element once
(the autosize code checks this), and we were applying it to the original
in the inputTemplate, so it wasn't being applied when we actually
instantiated a copy.

To fix it, wound up added the :notUnder() selector, which was more or
less suggested on SO.