Get the edited value through to UserSession
Checkpoint.  We are currently dropping the UserValue on the floor, but
that's better than the exceptions we were producing before.

Sadly, requires forwarding the wrapper to UserSession for now, so that
it has the info to forward to Space.  I expect this to eventually go
away, but that'll be a while yet.