Summaries are starting to work
Not pretty yet, but we are now successfully summarizing multiple users.

Introduced the SpaceAPI and SpacePlugin mechanisms, so that we can do
things like this inside Space processing without clogging up Space with
other systems' code.

Lots of tweaking, to get the Summary code working and the type
signatures all lined up.

Simplified the signature of modifyThing(), since it is now exposed to
SpacePlugins.

Introduced QLog.warn.  Experimentally gave that an Option[T] signature,
so that it can be used in the very common case where I want to print an
error when an Option unexpectedly returns None inside of a for
comprehension.