Got the new External Link Type fully working
Rewrote this as a Model Type, which is what I should have done in the
first place.

Wound up breaking it out into a separate Ecot, due to initialization
dependency issues.  (Link is too central to hold a type that is so
high-level.)  Move _withParam into that, since it is fundamentally about
External Links.  This simplifies the highly-depended-upon Links, which
is nice.

Along the way, split Links' MOIDs into PublicMOIDs vs. MOIDs, which I
think is probably a convention I need to adopt more broadly.  If a MOID
isn't used externally, don't expose it.