Critical fix for Model Values during Create Instance
Remarkably, it seems like using Model Values in Create Instance in the
old-style Editor has probably never worked correctly: some of the fields
in the Model Property would have been lost.  Surprised that I didn't hit
this before, but it seems to have only happened in this somewhat
specific situation.