_edit can now walk up the Context chain
Actually a huge (if mechanical) code change to get to this point.
Needed to rewrite the renderInput() powertrain to pass the QLContext
throughout instead of the RequestContext.  I've been expecting this
change eventually, but this is the straw that broke the camel's back.

Basically all changes were mechanical except for
Editor.getInstanceEditor(), which has the tiny but critical change to
pass in the existing context.

Introduced PropertyBundle.thisAsQValue, since I often don't need to
build a full QLContext (and shouldn't, since I want to use
context.next()).