Fixed _bulleted to not have excessive <ul> nesting
That is, every bullet item was winding up in its own <ul>, which was
certainly wrong.

Fix actually involved a fairly serious and long-standing bug in the *
""asCollection"" processing.  If an outer Text turned this on, it did
not get turned off for inner Texts, due to the way Context was being
passed down.  Had to introduce a pro-active concept of turning *off*
asCollection in order to fix that.