Added the _photoTarget function
This shows the first picture in a list, and serves as a target for
displaying a thumbnail full-sized.

To help this work, enhanced QLContext to keep track of the Property /
Function that was used to produce a given QValue.  This lets us look
back up the tree, and hook the target and the thumbnails together by the
property from whence they originated.