.3y285u6: _createButton now opens a new page
While having _createButton operate inline is consistent with the Bulk
Editor, it proves to rarely be what I actually want.  So added a new
entry point for creating a Thing, creating first and *then* showing the
Thing Editor, to be more consistent with the Thing Editor's modus
operandi, and repointed _createButton to that.

Changed a whole bunch of places where we are computing the link to
create a new Thing.  Those should probably be refactored together?