.3y285y7: Introduced the Choice Order Property
When linking to a Choice, if Choice Order is set on the Model, this
controls the order of the choices.  (Instead of listing them
alphabetically, which is the default.)

Note that Choice Order is on the Model, because it is somewhat common to
derive multiple Properties from a single Model.  So the Model defines
the Choice.