Genericized the Ecology
Not there yet, but seems to be compiling and working again.  Big
refactor of the Ecology, to separate the "generic" concepts from the
Querki-server specific ones.  Very hairy stuff.

Most important: I believe the "generic" parts (mostly) no longer rely on
TypeTag.  So there is a chance this might mostly work on the client.

Needed to add a bunch of apparently-spurious "with EcologyMember"s to
several Ecots to get them to work.  This requires further thought.

Note that SpaceState is a weird special case: had to change its
signature to fix some recursive type loops, since the main Ecology
classes depend on it.