We now fetch the rendered Thing via AJAX
This is a hair slower, but gets to first user response quicker, and is
more flexible for the future.

Introduced WrapperDiv, as a way to define a placeholder whose contents
can be replaced when AJAX completes.

Had to fix a bug in Gadget, which was preventing Gadget.render from
being called.