Removed the old Attachments mechanism
The attachment() entry point was using the deprecated (and now deleted)
getIdentityByThingId(), so I decided to bite the bullet and scrag the
entire Attachments concept, since I don't think we want to do that any
more.

This turns out to let me delete an *amazing* amount of code, scattered
all across the system.  Really, it simplifies the architecture, since it
removes one of the five Kinds.  This is a truly amazing win, I think.

Also, rewrote withSpace() to use the new getOwnerIdentity() instead of
the old getIdentityByThingId().  This means we are no longer doing
synchronous DB lookups for the owner's Identity by handle when loading a
page.  Also yay!