TagSetInputTest coming closer to working
Turns out that, for utest to work, we need to use the runNow
ExecutionContext.  Switched to that as the default for now.

Introduced a mechanism whereby tests can inject API entry points when
necessary.

Stubbed .manifest(), and the entry point for MarcoPolo, to make
TagSetInput not choke.

It turns out that async/await don't play nicely with synchronous errors
thrown by autowire's call().  So moved away from that for now.