Rewrites towards single-page model
MAJOR CHANGE: decided to bite the bullet and just call the Client a
Client: once you load it, it stays on the same page (changing hashes)
until you change Spaces.  This is a crucial change, but should speed
things up massively.

Lots of rewrites to that end, mainly to remove passing the Thing in at
Client setup, and instead have it be fetched dynamically after start.

Changed the URL, to occupy what used to be the URL for the Space itself.
That mechanism is long-since deprecated, so I removed it and changed all
references to it.