Got the standard Thing Edit button working
Almost disturbingly easy.  I love it when a plan comes together.  Does
need further refinement of QLButtonGadget to deal with code-defined
buttons, though.

IMPORTANT: also added the refinement of the Space in UserSpaceSession so
that you simply can't see Things you don't have Read access to.  This is
inefficient, but necessary for safety.  Make it better!