Got the Thing input working in ExplorePage
The problem was hooking: to work correctly, the MarcoPoloInput has to be
hooked *after* it gets added to its parent.  Invented a new mechanism so
that this happens consistently after page load.

Fixed some problems with MarcoPolo, which mostly boil down to it having
the same annoying type quirks as Manifest.

Enhanced MarcoPoloInput to take additional Modifiers.

Enhanced the MarcoPolo back end to cope with property-less queries.