Beginning to display Dialogs
Checkpoint towards Delete Thing.  Took a good deal of struggle to get
here.

Beginnings of the jQuery UI Facade.  Put this under org.scalajs, which
is probably where it belongs.  (??)

Had to fix the Manifest file, which turned out to include an out-of-date
and conflicting version of jQuery UI, which was causing things to fail.

Note that, for the moment, we are using the non-minified version of
jQuery UI.

Fixed things to use the built-in undefined variable, now that Sebastien
pointed out where it lives.