You can now reorder a List
This now works as a true reorder operation, instead of completely
rewriting the whole thing as it used to be.

This is the camel's nose in the tent of true *change* operations,
instead of complete rewrites the way we're been doing them heretofore.
These changes are a bit fragile at the moment, until we fix the
protocol, but fine for small spaces without contention.