You can now change a single List element
This turned out to have not been working; indeed, it wasn't previously
possible, since the infrastructure only allowed for parsing the entire
List.

The change is quite conservative, probably moreso than is ideal, but
this edit-handling stack really needs a complete makeover once the
Client is solid.

Also, added the SelectGadget, a simple InputGadget for select tags.