Fixed a crucial race condition in Add List Item
We were trying to send the new value before receiving confirmation that
the element had been created.  This was sometimes causing races where,
on the server side, we would try to set the value before creation.

To fix this, save() now exposes a Future.

Also, experimentally began using a couple of the new jQuery signatures
that I am planning on PR'ing soon.