Making the ThingInfo types more consistent
Also, fixed a horrible bug caused by the way the rc variable is
essentially a side-effect in the autowire handlers.  This workaround
gets it functioning again, but it's only an accident that we haven't hit
this before.

Progress towards the PropDetails pane.

Introduced spewing() into global on the server side.