Fixed a *horrible* bug that was preventing descriptions from working right
The issue was that, if you closed a PropertyDetails and then re-opened
it, it would show up empty.  After sweating about four hours of blood,
sweat and tears over this, finally figured out that the issue was
because Gadget itself was storing the render as a lazy val, so
re-rendering would wind up with that same value every time.

Also, beginnings of the Property Editor.