Several tweaks
One commit, because GfW is acting up again.

Refactored a bunch of classes out of ModelDesignerPage, to keep Eclipse
happier.

Workaround for the problem where a PropertySection is empty, so we
create a null entry as a fake drag target.

Fix for a bug in QTextUtils; code had fallen behind the times.

Fixed a bug in the subtitle of ThingPage.