Made API requests properly POST-y
Finally remembered that *anything* that appears in the routes file winds
up as a Query Param, even if it should obviously be a Form Data param.
Play forces me to be much more explicit about Form parameters.  So
switched the Client APIs to do that right.

This should fix a bug wherein submitting an especially long Large Text
change simply fails, because of an overlarge URL.