Fixes to security holes
Most importantly, fixed a bug in UserSpaceSession that was allowing
users who didn't have read permission to read a Thing.  Fixing this
causes it to work as expected: the Thing totally doesn't exist for them.
(Which will cause other issues, but is nice and secure.)

Also, added a mechanism at the front so that, if you can't read a Space
in the first place, the Client doesn't even admit that that Space
exists.