Fixed a bug that was showing errors after login
The problem was that _userApiRequest was calling withUser(true).  So if
an Anonymous user, using the Client, looked at a page, it would wind up
calling this, and thence onUnauthorized, which would then set returnTo
to _userApiRequest.  After you logged in, it would try to redirect to
there -- which is nonsensical, because it's an API entry point, not a
page.