Get joinSpace working again
This had gotten scragged by the security changes, and is a good example
of the danger of getting too casual about sending the Space all over the
place.  Since the Space it was received is a UserSpaceSession version,
and of course this non-Member can't see the Person records,
acceptInvite() simply crashed and burned.

So rewrote this properly, moving the key logic into a new Actor in the
Space's Troupe.