Fixed an idiotic bug that I introduced in the name of "optimization"
That will teach me not to doubt when my instincts are screaming, "This
is a risky hack".

The caching of the model's pointer made a little sense in isolation, but
didn't take into account the fact that the Model Thing itself can then
get edited, and we are left with a pointer to an obsolete Model.  Duh.