Fixed Model Values to not get over-initialized
This fixes a subtle but nasty bug: Model Values were being initialized
(via doDefault) with *copies* of the Properties from the Model.  Which
means that changes in the Model get blocked by the local copy in the
Property.

Simply removing that fixes the problem: the Model Value gets
pass-throughs for all of the Model's Properties, so it looks the same to
the user but only gets local copies when you make a change.