Uploading photos from files is now working
Current code is a semi-literal translation of the old Javascript.  I
suspect we'll gradually want to improve on that, but it's a start.

Taking a photo from my phone does *not* appear to work yet.

Had to upgrade to scala-js-dom 0.6.1 to get the corrected version of the
FileReader class.

Includes partial facade for the FileUpload library.

Fixed a bug in Pages.showSpacePage().