Added the Make a Page button to QL Explorer
Getting this right was a little subtle.  Turned out that we were firing
the events from page load too early, before it had fully rendered.

Enhanced showPage to return a Future, which is just plain useful.

Added a horrible hack to deal with the fact that the path in the name of
existing control is old-fashioned, using OIDs instead of TIDs.