Broadened the image-upload to work well with PNG
As it turns out, fileupload automatically sets the header of the stream
to the actual image type, so most of the problem was propagating that
and adjusting accordingly.

Note that this is not fully generalized yet: it really just supports PNG
and JPEG.  But adding new formats should be pretty easy if they are
supported by Java.

Note that PNGs are *not* auto-sized; the presumption is that you want
them full-size.