Weakened the change to force Name
My earlier change to make the Name property show up in the Advanced
Editor if Derive Name isn't checked turns out to break the unit tests:
Name shows up in _foreachProperty despite not being in the Instance
Properties.  And the thing is, it probably shouldn't do so there.  So
hacked things even worse...