Refactored SpaceRegion out of SpaceManager.
SpaceRegion is essentially what will become a Cluster Sharding Region pretty soon, but for now let's separate it from the code that *can't* be done with Cluster Sharding.

This is a fairly small change with *incredibly* dramatic consequences. By separating SpaceMessage from SpaceMgrMessage, we touch an awful lot of code.

So I finally decided to begin removing the deprecated old-UI code. This is crossing the rubicon in a big way: I'm beginning to delete dead code wholesale. Hopefully I won't regret that.