IMPORTANT: made Ecology a static global again
Finally decided to let ease of coding win out over purity. Passing the Ecology around into everything has been an ongoing headache, and it's about to become a much bigger one when we get into clustering, because the Ecology pointer absolutely can't serialize.

So in the name of not making myself pointlessly crazy, decided to back off to the impure but much easier approach, where the Ecology is a global. This *enormously* simplifies an awful lot of code.