Begin to support namespaces
Pretty simplistic for the moment, but may be enough for the time being -- all we really need to do is distinguish stuff in SystemSpace from user-defined. We should be able to easily enhance this later to distinguish what comes from Apps as well. (That's why we have prefixMap instead of just hard-coding SystemSpace.)