Checkpoint: starting to refactor out Notifications
What do Notification have to do with it? Well, ImportSpace really wants to be under an AutowireApiImpl within UserSession, but UserSession's Autowire handling is way behind the times. So I need to make NotificationFunctionsImpl more normal.

New concept: the ImplCache is there so that Impls can cache data as needed. This should allow the NotificationFunctionsImpl to absorb the code currently in UserNotifications.