XML fix and UTF-8 fix
Improved the XML parser to cope with some of the common XML Entities that get used in export/import. Not comprehensive yet, but a decent start.

Fixed all the DB table-creation code to explicitly specify UTF-8. The dev machine turns out to default to latin1, and that breaks Unicode horribly. Let's not leave this up to installation happenstance.

These two fixes together get us much closer to correct export/import on most Things.