Switching to a CrossProject.Full for the shared code
This is necessary so that I can do slightly different things from the JVM and JS code. Note that, while an awful lot of files changed, most of them just moved location.

Had to make some subtle tweaks to deal with the Html class. Decided to turn it into a real class on both sides, and rename it to QHtml, since just reusing the Twirl Html class on the JVM side is a little semantically dicey. Note that the shared code needs to include the SbtTwirl plugin (but specifically *not* the PlayScala plugin) in order to access the Html class.