Introduce the apps parameter to SpaceState
Split apps and system: I think treating System as a special case is reasonable. (Why do it? So that System is always at the end of the search, despite it being depth-first.)

Added some higher-level utility functions, so that application code mostly doesn't need to think about the details.

Passes unit tests, but still needs smoketesting.