Remove the ancient and idiotic indirection from PropFetcher to PropMap
Realized that this is becoming crucial, because PropFetcher isn't serializable.

Honestly have no idea any more *why* I wrote it this way in the first place. Has to date from the earliest days of Querki. Doesn't appear to be relevant any more, though, which leads me to suspect it was an ill-considered and now redundant workaround for some problem.