_QLTree starting to output the right stuff
Most of the changes aren't relevant. Turned out that the QL interface was being pulled into CoreEcot, which is *wildly* inappropriate coupling. This was causing problems, because I need it as an initRequires, and it was always there as an interface. So pulled it out of CoreEcot, and added it where needed.

The more substantive change was the addition of the new processAsOpt and rawParam methods to Invocation.